docs working as expected
